fix : refactor
