+++ src/frontend/screen/index.js
@@ -4,7 +4,7 @@
 
 
 export { Home } from "./Home/Home";
-export { Signin } from "./auth/Signin";
-export { Signup } from "./auth/Signup";
+export { Signin } from "./Signin";
+export { Signup } from "./Signup";
 export { Explore } from "./Explore/Explore";
 export { NotFound } from "./NotFound/NotFound";
